requirements.txt
aiogram==3.2.0
aiofiles==23.2.1
python-dotenv==1.0.0
aiosqlite==0.19.0
requests==2.31.0
aiohttp==3.9.1
feedparser==6.0.10
openai>=1.13.3
httpx>=0.27.0
telethon==1.31.1
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3>=2.0.7 